+++ src/interfaces/molden.jl
@@ -9,6 +9,7 @@ using AtomsBase
 using Printf
 using ..ElemCo.Utils
 using ..ElemCo.ECInfos
+using ..ElemCo.QMTensors
 using ..ElemCo.MSystem
 using ..ElemCo.BasisSets
 using ..ElemCo.Wavefunctions
@@ -105,7 +106,7 @@ function write_molden_orbitals(EC::ECInfo, filename::String)
       maxl > 5 && println(f, "[13I]")
     end
     #TODO use correct energies and occupations
-    if is_restricted_MO(orbs)
+    if is_restricted(orbs)
       energies = zeros(size(orbs,2))
       occupation = zeros(size(orbs,2))
       printmos(f, orbs, order, energies, occupation)